manager: Generate random super key to instead of the default one
It's not safe to set a default superkey to a common user, it will be leaked and some apps can be detected APatch with the default superkey.

We have introduce random uuid as the random superkey to instead of the default key, and it will show to user at the end of flash, there's nobody can get it except user.

But it'sobviously, if APatch user forget to record it or forget it, there's no way to get it unless APatch leaked the local superkey.

Test: Build Manager - Flash in Third-Party Recovery - Record the random key - Reboot - Enter the key - Check APatch

Result: It works well(Tested on Oneplus Ace 3 Pro)

Signed-off-by: SakuraKyuo <SakuraKyuo@protonmail.com>