script: Support flash APatch in AOSP Based Recovery
After Android 11, the AOSP-based Recovery no longer has /sbin, while TWRP and LineageOS-based Recovery retain the /sbin/sh interpreter to provide compatibility with this shell flashing script. TWRP also adds bash to /sbin.

In order to adapt to AOSP-Based Recovery, we canceled the use of /sbin/bash to run the installation script, and instead used APatch's built-in busybox to run the installation script to achieve the purpose of adapting to all recovery.

We don't need updater-script anymore, so keep it empty. The core part is in update-binary and addon folder.

Test: Build Manager - Flash in AOSP-Based Recovery - Reboot - Enter the key - Check APatch

Result: It works on AOSP-Based & LineageOS-Based Recovery now, TWRP also works well(Tested on Xiaomi MI MIX & Oneplus Ace 3 Pro)

Signed-off-by: SakuraKyuo <SakuraKyuo@protonmail.com>