script: Switch to a more reliable way to get log
Previously, we used to output logs directly to /dev/tmp/install/log, but after multiple rounds of testing, this method only obtains the previous line log of the core error, which brings inconvenience to development work troubleshooting, so we changed the acquisition method.

However, due to unknown reasons, the core error may be output at the beginning of the log output (i.e. the first line). Therefore, when the user feedback patch failed, we need to ask for the user to provide a screenshot of the starting position to determine the core error.

Test: Build Manager - Flash on a device that can't patch APatch normally - Check output in the screen

Result: It will print the full log with the core error now(Tested on Xiaomi MI MIX)

Signed-off-by: SakuraKyuo <SakuraKyuo@protonmail.com>