script: Add kernel flag detection
According to the APatch documentation, the APatch patch requires setting the kernel flag CONFIG_KALLSYMS to y (i.e. enabling the kernel flag). kptools will report an error "analyze_kallsyms_info error" and exit because CONFIG_KALLSYMS is not enabled.

In order to intercept this problem at the initial stage of installation and reduce the workload of developers, extract-ikconfig provided by https://github.com/torvalds/linux was added. The flashing script will obtain whether the kernel flag is turned on. If it is turned off, the installation script will abort and return an error message to the user.

Test: Build Manager - Flash on a device that not enable the flag - Check output in the screen

Result: It will abort and print the reason in the screen(Tested on Xiaomi MI MIX)

Signed-off-by: SakuraKyuo <SakuraKyuo@protonmail.com>