manager: Add kernel configuration detection when patching in APP
Same as the commit https://github.com/SakuraKyuo/APatch/commit/626b9124c66050efb28c352c3d01d7f6410c474c, according to the documentation, enabling this kernel configuration is required. So the same detection method as the commit is used to detect kernel configuration when using apk to patch the kernel.

At the same time, because the grep command provided by busybox lacks the parameters we need, the grep in the toybox provided by the system is also used to replace the grep in busybox to make the grep command in extract-ikconfig work properly.

Test: Build Manager - Patch the kernel that not enable the flag configuration - Check output

Result: It will abort when it NOT detected the kernel configuration.

Signed-off-by: SakuraKyuo <SakuraKyuo@protonmail.com>