scripts: Fixed detecting kernel configuration in APP
In commit https://github.com/bmax121/APatch/commit/6d62eaebcf6fdfa08253da2386d2b17557bbce24, we used mktemp provided by toybox to avoid the problem of missing mktemp. However, the test found that permission denied would occur on some devices because mktemp provided by toybox would create temporary files in /tmp, which is not possible in non-Recovery. Therefore, we added the parameter of passing in the APP cache directory to the APP patch script as the directory for mktemp to generate temporary files. This parameter is not passed in when flashing in Recovery mode.

Test: Build Manager - Patch in the device that not have mktemp - Check result

Result: It's working fine now(Tested on Redmi Note 11 5G)

Signed-off-by: SakuraKyuo <SakuraKyuo@protonmail.com>